Flutter fix-1: analyze errors resolved (from user's first real run)
- TranslateLanguage.fromBcp (galat API — 3 errors) → apna mlLang() bcpCode lookup
  helper (home.dart + info.dart fix)
- test/marine_test.dart — REAL unit tests (haversine/bearing/xtrack/steer/eta/
  compass16, reviewer-verified values) — flutter create ka widget_test.dart
  delete karne ko README mein likha (wo default counter-app ka hai)
- warnings sweep: _reloadGps removed, 2× unused ThemeData vars, unnecessary cast,
  unnecessary fix!, _zoom → final
- README: AndroidManifest permissions block (INTERNET/FINE+COARSE LOC/SEND_SMS/
  POST_NOTIFICATIONS/VIBRATE/WAKE_LOCK) + telephony-discontinued note
- verified in sandbox: balance 0 issues, tr-key coverage complete

+++ orca_flutter/lib/state.dart
@@ -65,7 +65,7 @@ class Settings extends ChangeNotifier {
 
   Future<void> zoom(bool up) async {
     final next = (scale + (up ? 0.15 : -0.15)).clamp(0.85, 1.75);
-    scale = double.parse((next as double).toStringAsFixed(2));
+    scale = double.parse(next.toStringAsFixed(2));
     notifyListeners();
     (await SharedPreferences.getInstance()).setDouble(_kScale, scale);
   }
